% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{quartorsv_cat_pandoc_fig}
\alias{quartorsv_cat_pandoc_fig}
\title{Cat pandoc command to include a figure}
\usage{
quartorsv_cat_pandoc_fig(
  path,
  cap = NULL,
  label = NULL,
  args = NULL,
  force_cross_ref = TRUE,
  force_exists = TRUE
)
}
\arguments{
\item{path}{character. Path to the figure.}

\item{cap}{character. Caption of the figure.}

\item{label}{character. Label of the figure.
Need not supply \verb{#fig-} prefix (but it
will not be repeated if you do).}

\item{args}{character. Additional arguments to be passed to pandoc.}

\item{force_cross_ref}{logical.
If \code{TRUE}, then will throw an error if either \code{cap}
or \code{label} are \code{NULL} as these are both required for
cross-referenceable figures in \code{quarto}.
Default is \code{TRUE}.}

\item{force_exists}{logical.
Will throw an error if the file at \code{path} does not exist.
Default is \code{TRUE}.}
}
\description{
Cat pandoc command to include a figure
}
\examples{
quartorsv_cat_pandoc_fig(
  path = "path/to/fig.png",
  cap = "Incredible hist() plot",
  label = "hist-plot",
  args = "width=50\%",
  force_exists = FALSE
)

}
